chore: update templates based on changes in upstream
We don't need duplicated manifest.yaml in each template anymore. Now
they just live in the language/runtime directory. The eslint and prettier
configs are duplicates so can be removed.

There is also a stray manifest.yaml in templates/typescript/cloudevents
that does not affect the upstream, but does have the effect of overriding
our ubi-based manifest.yaml and it is removed for now.

Signed-off-by: Lance Ball <lball@redhat.com>